Modify webpage elements to reflect desired outcome
Centered text positioning of header links and fixed the issue of text
wrapping after one word. Fixed the same issue on the links in the nav
bar and centered the links there with appropriate spacing as well.
Added spacing between the elements in the sidebar to match the desired
outcome of the assignment. Centered the header logo to match the
positioning of the header links. Further centered to logo on the footer.